# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is a OpenAPI 2 specification for Mashrooy"
  version: "1.0.0"
  title: "OpenAPI Mashrooy"
host: "localhost:3006"
basePath: "/Mashrooy"
...
schemes:
- "http"
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
  /app:
    get:
      tags:
      - "App"
      description: "list all layout configs"
      operationId: "getAppLayoutList"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/App"
      security:
      - api_key: []
      x-swagger-router-controller: "App"
    post:
      tags:
      - "App"
      summary: "create configuration for whole layout"
      description: "add new custom config for header,sidebar,footer,aside and styling"
      operationId: "addAppLayout"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "layout"
        description: "object of the new custom layout config"
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/App"
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/App"
      security:
      - api_key: []
      x-swagger-router-controller: "App"
  /app/{id}:
    get:
      tags:
      - "App"
      summary: "get configuration for whole layout"
      description: "get custom config for header,sidebar,footer,aside and styling"
      operationId: "appGetOne"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of item that needs to be fetched"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/App"
      security:
      - api_key: []
      x-swagger-router-controller: "App"
    put:
      tags:
      - "App"
      summary: "update configuration for whole layout"
      description: "update custom config for header,sidebar,footer,aside and styling"
      operationId: "updateAppLayout"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of item that needs to be fetched"
        required: true
        type: "string"
      - in: "body"
        name: "layout"
        description: "object of the new custom layout config"
        required: true
        schema:
          $ref: "#/definitions/App"
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/App"
      security:
      - api_key: []
      x-swagger-router-controller: "App"
    delete:
      tags:
      - "App"
      summary: "delete configuration for whole layout"
      description: "delete custom config for header,sidebar,footer,aside and styling"
      operationId: "deleteAppLayout"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of item that needs to be deleted"
        required: true
        type: "string"
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/App"
      security:
      - api_key: []
      x-swagger-router-controller: "App"
...
  /branch:
    get:
      tags:
      - "Branch"
      summary: "List All branches"
      description: ""
      operationId: "listBranches"
      parameters:
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Branch"
        500:
          description: "Internal Server Error"
          schema:
            type: "string"
      security:
      - loanstore_auth:
        - "write:loans"
        - "read:loans"
      x-swagger-router-controller: "Branch"
    post:
      tags:
        - Branch
      summary: adds branch
      description: Adds branch to the system
      operationId: createbranch
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: branches
          description: branches to add
          required: true
          schema:
            type: array
            items:
              $ref: '#/definitions/Branch'
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Branch'
        '400':
          description: Validation Error
          schema:
            $ref: '#/definitions/ApiResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        '409':
          description: Conflict record
          schema:
            $ref: '#/definitions/ApiResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      x-swagger-router-controller: Branch
    put:
      tags:
        - Branch
      summary: Update an existing Branchs
      description: ''
      operationId: updateBranchs
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: Branch object that needs to be updated
          required: true
          schema:
            type: array
            items:
              $ref: '#/definitions/Branch'
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Branch'
        '400':
          description: Validation Error
          schema:
            $ref: '#/definitions/ApiResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        '404':
          description: Record Not Found
          schema:
            $ref: '#/definitions/ApiResponse'
        '409':
          description: Conflict record
          schema:
            $ref: '#/definitions/ApiResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      x-swagger-router-controller: Branch
    delete:
      tags:
        - Branch
      summary: Delete branches By Id
      operationId: deleteBranchesById
      parameters:
        - in: body
          name: IDs
          description: array of IDs to be deleted
          required: true
          schema:
            type: array
            items:
              $ref: '#/definitions/Branch'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/ApiResponse'
        '400':
          description: Validation Error
          schema:
            $ref: '#/definitions/ApiResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      x-swagger-router-controller: Branch
  /branch/listByGeo:
    get:
      tags:
      - "Branch"
      description: "Get  list of branchs by GEO"
      operationId: "getBranchByGeo"
      parameters:
      - name: "longitude"
        in: "query"
        description: "longitude of the area that needs to fetch own branches"
        required: true
        type: "string"
      - name: "latitude"
        in: "query"
        description: "latitude of the area that needs to fetch own branches"
        required: true
        type: "string"
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Branch"
        500:
          description: "Internal Server Error"
          schema:
            type: "string"
      security:
      - loanstore_auth:
        - "write:loans"
        - "read:loans"
      x-swagger-router-controller: "Branch"
...
securityDefinitions:
  loanstore_auth:
    type: "oauth2"
    authorizationUrl: "https://tasaheel.swagger.io/oauth/authorize"
    flow: "implicit"
    scopes:
      write:loans: "modify loans in your account"
      read:loans: "read your loans"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  App:
    type: "object"
    example:
      layout:
        aside:
          tabs:
          - icon: "icon-list"
            index: "1"
            value: "icon-list"
          - icon: "icon-list"
            index: "1"
            value: "icon-list"
          content:
          - index: "1"
            type: "notifList"
            desc: "Hello Buddy"
          - index: "1"
            type: "notifList"
            desc: "Hello Buddy"
        styling:
          typography:
            text: "1rem"
            title: "2rem"
            font: "'Open Sans', sans-serif"
          spacing: "1rem"
          breakpoints:
            md: "64em"
            sm: "30em"
            xs: "0em"
            lg: "75em"
          colors:
            secondary: "#7F8FA4"
            success: "#1fce6d"
            warning: "#f2c500"
            danger: "#e94b35"
            error: "#e94b35"
            primary: "#2c97de"
          base: "16px"
        breadcrumb: "{}"
        footer:
          company:
            name: "Room Art Space"
            url: "https://google.com"
          poweredBy:
            name: "Raseedy"
            url: "https://google.com"
        sidebar:
          nav:
            items:
            - badge:
                variant: "info"
                text: "NEW"
              name: "Dashboard"
              icon: "icon-speedometer"
              url: "/dashboard"
            - badge:
                variant: "info"
                text: "NEW"
              name: "Dashboard"
              icon: "icon-speedometer"
              url: "/dashboard"
          form: "{}"
          footer:
            desc: "you can win"
          header: "{}"
          minimizer: "{}"
        header:
          appNavbarBrand:
            minimized:
              logo:
                'src:': ""
                alt: ""
            full:
              logo:
                'src:': ""
                alt: ""
          primaryMenu:
          - name: "Dashboard"
            index: 0
            type: "NavItem"
            class: "nav-link"
            url: "/dashboard"
          - name: "Dashboard"
            index: 0
            type: "NavItem"
            class: "nav-link"
            url: "/dashboard"
          secondaryMenu:
          - badge:
              variant: "danger"
            name: "Dashboard"
            icon: "icon-bell"
            menuItems:
            - name: "You have new notifications"
              type: "header"
            - name: "You have new notifications"
              type: "header"
            index: 3
            type: "StyledNavItem"
            class: "nav-link"
            url: "/dashboard"
          - badge:
              variant: "danger"
            name: "Dashboard"
            icon: "icon-bell"
            menuItems:
            - name: "You have new notifications"
              type: "header"
            - name: "You have new notifications"
              type: "header"
            index: 3
            type: "StyledNavItem"
            class: "nav-link"
            url: "/dashboard"
        main:
          hasAside: true
          hasSidebar: true
...
  Branch:
    type: "object"
    properties:
      id:
        type: "string"
      branch_code:
        type: "integer"
      branch_name:
        type: "string"
      en_name:
        type: "string"
      districtId:
        type: "string"
      governantId:
        $ref: "#/definitions/Governorate"
      branch_add:
        type: "string"
      en_address:
        type: "string"
      geo:
        $ref: "#/definitions/Branch_geo"
      update_stamp:
        type: "string"
    example:
      branch_code: 0
      geo:
        latitude: "latitude"
        longitude: "longitude"
      districtId: "districtId"
      en_address: "en_address"
      branch_name: "branch_name"
      branch_add: "branch_add"
      en_name: "en_name"
      id: "id"
      governantId: "governantId"
      update_stamp: "update_stamp"
  Governorate:
    required:
    - "id"
    - "name"
    properties:
      id:
        type: "string"
        example: "d290f1ee-6c54-4b01-90e6-d701748f0851"
      name:
        type: "object"
        properties: 
           ar: 
             type: string
           en: 
             type: string
...
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"
...
  Branch_geo:
    type: "object"
    properties:
      longitude:
        type: "string"
      latitude:
        type: "string"
    example:
      latitude: "latitude"
      longitude: "longitude"
...